pyproject.toml
[tool.poetry]
authors = ["pykulytsky <pykulytsky@gmail.com>"]
description = ""
license = "MIT"
name = "mahi"
version = "0.1.0"

packages = [
  {include = "scripts", from = "."},
]

[tool.poetry.dependencies]
PyJWT = "1.7.1"
SQLAlchemy = "^1.4.39"
SQLAlchemy-Utils = "^0.38.3"
aiokafka = "^0.7.2"
aioredis = "^2.0.1"
alembic = "^1.8.0"
celery = "^5.2.7"
fastapi = "^0.78.0"
fastapi-plugins = "^0.10.0"
fastapi_permissions = "^0.2.7"
httpx = "^0.23.0"
passlib = "^1.7.4"
psycopg2 = "^2.9.3"
pydantic = {extras = ["email", "dotenv"], version = "^1.9.1"}
pylint = "^2.15.2"
python = "^3.10"
python-multipart = "^0.0.5"
requests = "^2.28.1"
sentry-sdk = "^1.7.0"
sqlmodel = "^0.0.8"
sse-starlette = "^1.1.6"
uvicorn = "^0.18.2"
websockets = "^10.3"
redis = {extras = ["hiredis"], version = "^4.3.4"}
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
black = "^22.6.0"
flake8 = "^4.0.1"
ipython = "^8.4.0"
isort = "^5.10.1"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
pytest-cov = "^3.0.0"
pytest-env = "^0.6.2"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.8.2"
pytest-picked = "^0.4.6"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry.scripts]
dev = 'scripts.dev:main'
prod = 'scripts.prod:main'

[tool.isort]
profile = "black"
